server api secured successfully. middleware removed until PostMan testing are complete
