+++ server.js
@@ -3,8 +3,9 @@ const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
 const routes = require("./routes");
 const cors = require('cors');
-const redisClient = require("./redis/main.js")
-
+const redisClient = require("./redis/main.js");
+const jwt = require('express-jwt');
+const jwks = require('jwks-rsa');
 //Yo Justin, can you figure out what add on is needed to get redis running in heroku??
 // if (process.env.NODE_ENV !== 'production') {
 //   redisClient.client.on("connect", function () {
